TRADE4ME_2.0/TRADE4ME_2.0/bot_pancakeswap_bsc/requirements.txt
web3==6.15.1
python-dotenv==1.0.1
requests==2.31.0
